requirements.txt
requests
click
pyjwt
-e .
